[IMP][10.0] contract: Add templating (#42)
Add template functionality for contracts
