+++ contract/README.rst
@@ -67,6 +67,7 @@ Contributors
 * Pedro M. Baeza <pedro.baeza@tecnativa.com>
 * Carlos Dauden <carlos.dauden@tecnativa.com>
 * Angel Moya <angel.moya@domatix.com>
+* Dave Lasley <dave@laslabs.com>
 
 Maintainer
 ----------
