+++ contract/__manifest__.py
@@ -4,19 +4,21 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 {
-    'name': 'Contracts Management recurring',
-    'version': '10.0.1.0.0',
+    'name': 'Contracts Management - Recurring',
+    'version': '10.0.1.0.1',
     'category': 'Contract Management',
     'license': 'AGPL-3',
-    'author': "OpenERP SA,"
-              "Tecnativa,"
+    'author': "OpenERP SA, "
+              "Tecnativa, "
+              "LasLabs, "
               "Odoo Community Association (OCA)",
     'website': 'https://github.com/oca/contract',
     'depends': ['base', 'account', 'analytic'],
     'data': [
         'security/ir.model.access.csv',
         'data/contract_cron.xml',
-        'views/contract.xml',
+        'views/account_analytic_account_view.xml',
+        'views/account_analytic_contract_view.xml',
         'views/account_invoice_view.xml',
     ],
     'installable': True,
