+++ contract/models/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
-# © 2016 Carlos Dauden <carlos.dauden@tecnativa.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import contract
-from . import invoice
+from . import account_analytic_contract
+from . import account_analytic_account
+from . import account_analytic_invoice_line
+from . import account_invoice
